telegram_bot_webhook_fixed_final/requirements.txt
python-telegram-bot[webhooks]==20.7
flask==2.3.3
google-api-python-client==2.120.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0